fixes to packeges
