--- config.json
@@ -1,5 +1,0 @@
-{
-  "couchdb": {
-    "uuid": "e5fb794f-2a50-484e-99fe-3cf79836f267"
-  }
-}
