+++ package.json
@@ -14,13 +14,12 @@
         "@hapi/iron": "^7.0.0",
         "@magic-sdk/admin": "^1.4.1",
         "@sendgrid/mail": "^7.6.2",
-        "@types/pouchdb": "^6.4.0",
         "cors": "^2.8.5",
         "dotenv": "^16.0.0",
         "ethers": "5.5.1",
         "iron-session": "^6.1.2",
-        "magic-sdk": "^8.1.1",
-        "nano": "^10.0.0",
+        "magic-sdk": "^10.0.0",
+        "nano": "^10.1.0",
         "next": "^12.1.7-canary.6",
         "nextjs-cors": "^2.1.1",
         "react": "17.0.2",
@@ -28,19 +27,11 @@
         "react-dom": "17.0.2",
         "siwe": "^1.1.6",
         "swr": "^1.3.0",
-        "wagmi": "^0.5.9"
+        "wagmi": "^0.7.12"
     },
     "devDependencies": {
-        "@types/cors": "^2.8.12",
-        "@types/node": "17.0.24",
-        "@types/react": "18.0.5",
-        "@types/react-datepicker": "^4.4.2",
-        "@types/react-dom": "18.0.0",
-        "autoprefixer": "^10.4.7",
-        "eslint": "8.13.0",
-        "eslint-config-next": "12.1.5",
-        "postcss": "^8.4.13",
-        "tailwindcss": "^3.0.24",
-        "typescript": "4.6.3"
+        "@types/react": "18.0.25",
+        "@types/react-datepicker": "^4.8.0",
+        "typescript": "4.8.4"
     }
 }
