+++ pages/_app.tsx
@@ -15,12 +15,9 @@ import { InjectedConnector } from 'wagmi/connectors/injected'
 import { MetaMaskConnector } from 'wagmi/connectors/metaMask'
 import { WalletConnectConnector } from 'wagmi/connectors/walletConnect'
 
-const alchemyId = process.env.ALCHEMY_ID
-
 // Configure chains & providers with the Alchemy provider.
 // Two popular providers are Alchemy (alchemy.com) and Infura (infura.io)
 const { chains, provider, webSocketProvider } = configureChains(defaultChains, [
-  alchemyProvider({ alchemyId }),
   publicProvider(),
 ])
 
@@ -28,7 +25,7 @@ const { chains, provider, webSocketProvider } = configureChains(defaultChains, [
 const client = createClient({
   autoConnect: true,
   connectors: [
-    // new MetaMaskConnector({ chains }),
+    new MetaMaskConnector({ chains }),
     new CoinbaseWalletConnector({
       chains,
       options: {
@@ -52,7 +49,6 @@ const client = createClient({
   provider,
   webSocketProvider,
 })
-
 //@ts-ignore
 const MyApp = ({ Component, pageProps, auth }) => {
   return (
